requirements.txt
streamlit
numpy
Pillow
torch
segment-anything
streamlit-image-coordinates
nibabel
opencv-python
tqdm
scipy
matplotlib
